Optimize function draw_staff
Formerly, its third param was a two-bit number indicating whether a high or low
ledger line was needed; from that param we calculated a bit string indicating
whether each row was a line or a space. Now, we pass in that bit string
directly; most callers pass a constant.

Additionally, a benign warning about non-initialized array cs and cL is fixed.