Support slim barlines
Slim barlines don't have the usual blank columns before and after them.
They are useful when two notes are tied across a barline.

Added four slim barlines:
  Single, slim
  Double, slim
  Double with left repeat, slim
  Double with right repeat, slim

The preexisting double wide barline doesn't end in a blank column because it's
expected to go at the very end of a composition. It does start with a blank
column, and the new double slim barline does not.

The preexisting double with left and right repeats barline was already slim.
A wide version would take six characters, which isn't supported.