Change note encoding to improve the options for beamed notes
This is the second of two changes that I noted in commit 01687fe I was planning
to make.

The old byte group type
  Note/Rest (2 bytes)
is split in two:
  Note or rest, not beamed (2 bytes)
  Note, beamed (3 bytes)

These two byte group types share the following slices:
  Bits 1-3:  Prefix indicating the byte group type
  Bits 4-5:  Accidentals
  Bits 6-7:  Articulations
  Bit  8:    Tie after note
  Bits 9-12: Pitches
  Bit  16:   Dotted
They differ in these slices:
  Bits 13-15, containing options for note duration and appearance
  Bits 17-24, the third byte that only "Note, beamed" has, containing options
              for beams

The new "Note or rest, not beamed" byte group type is similar to the old
"Note/Rest" byte group type, but they are not backwards-compatible. Most of the
same slices are available, except that "Note or rest, not beamed" does not
support beams. (This freed up one bit, which I used to expand the prefix slice
from two bits to three.) I have reordered the slices as described above. This
was necessary to give "Note, beamed" a reasonable order of slices and give
"Note or rest, not beamed" the same order of shared slices.

The benefits of this enhancement are:
1.  Formerly, a new beamed note's number of left beams (1 or 2) was determined
    by the previous beamed note's number of right beam. However, music notation
    commonly contains a two-beamed note connected to a one-beamed note and vice
    versa. This is now supported.
2.  Formerly, there were only two options for stem height of beamed notes
    (2-3). Now, there are four options (1-4). This enables creating a beam
    between notes more than one pitch apart.
3.  The duration, flag, and beam encoding options are easier for a user to
    understand.